requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.2
colorama==0.4.6
ecdsa==0.19.2
fastapi==0.136.0
greenlet==3.4.0
h11==0.16.0
idna==3.11
passlib==1.7.4
psycopg2==2.9.12
pyasn1==0.6.3
pydantic==2.13.3
pydantic_core==2.46.3
python-jose==3.5.0
rsa==4.9.1
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
uvicorn==0.45.0
python-dotenv==1.2.2
bcrypt==4.0.1
websockets==16.0
python-multipart==0.0.26